Add ensemble ASE calculator
