+++ example/predict/run_ase.py
@@ -2,12 +2,19 @@
 from newtonnet.utils.ase_interface import MLAseCalculator
 from ase import Atoms
 
-mlcalculator = MLAseCalculator(model_path='training_18/models/best_model_state.tar',
-                               settings_path='training_18/run_scripts/config0.yml',
-                               method='fwd_diff',    # 'autograd', 'fwd_diff', or None
-                               grad_precision=0.0001,    # optional, grid size in Angstrom for Hessian gradient calculation (default: 0.0001)
-                               device='cpu'   # 'cpu' or list of cuda
-               )
+mlcalculator = MLAseCalculator(
+    model_path=['training_8/models/best_model_state.tar',
+                'training_9/models/best_model_state.tar',
+                'training_10/models/best_model_state.tar',
+                'training_11/models/best_model_state.tar'],
+    settings_path=['training_8/run_scripts/config.yml',
+                   'training_9/run_scripts/config.yml',
+                   'training_10/run_scripts/config.yml',
+                   'training_11/run_scripts/config.yml'],
+    method='autograd',    # 'autograd', 'fwd_diff', or None
+    #grad_precision=0.0001,    # optional, grid size in Angstrom for Hessian gradient calculation (default: 0.0001)
+    device='cpu'   # 'cpu' or list of cuda
+)
 
 h2 = Atoms('H2',
            positions=[[0.0, 0.0, 0.0],
