+++ example/predict/training_10/log.csv
@@ -1,0 +1,39 @@
+epoch,loss(MSE),lr,time,tr_E(MAE),tr_F(MAE),val_E(MAE),val_F(MAE),irc_E(MAE),irc_F(MAE),test_E(MAE),test_F(MAE)
+1.00000,153.72802,0.00010,13424.46260,3.00583,1.89310,2.15493,1.54842,0.00000,0.00000,0.00000,0.00000
+2.00000,76.92717,0.00010,13417.06655,1.92707,1.45095,1.89547,1.34538,0.00000,0.00000,0.00000,0.00000
+3.00000,61.39961,0.00010,13359.31836,1.67265,1.31332,1.58093,1.28859,0.00000,0.00000,0.00000,0.00000
+4.00000,53.38987,0.00010,13387.80810,1.55332,1.23209,1.34426,1.23222,0.00000,0.00000,0.00000,0.00000
+5.00000,48.39023,0.00010,13394.40938,1.48377,1.17730,1.61153,1.19224,0.00000,0.00000,0.00000,0.00000
+6.00000,44.91737,0.00010,13384.59401,1.42335,1.13701,1.47478,1.17611,0.00000,0.00000,0.00000,0.00000
+7.00000,42.22292,0.00010,13386.59971,1.38673,1.10412,1.42443,1.07291,0.00000,0.00000,0.00000,0.00000
+8.00000,40.24889,0.00010,13397.96076,1.35468,1.07934,1.98090,1.05749,0.00000,0.00000,0.00000,0.00000
+9.00000,38.63180,0.00010,13441.67106,1.32959,1.05809,1.41160,1.09611,0.00000,0.00000,0.00000,0.00000
+10.00000,37.37408,0.00010,13394.48903,1.32121,1.04061,1.52590,1.07899,0.00000,0.00000,0.00000,0.00000
+11.00000,36.25526,0.00010,14344.56433,1.29887,1.02566,1.61051,1.02508,0.00000,0.00000,1.61050,1.02407
+12.00000,35.32537,0.00010,13391.15156,1.28603,1.01261,1.17257,1.01512,0.00000,0.00000,0.00000,0.00000
+13.00000,34.49950,0.00010,13382.29707,1.27315,1.00045,1.31551,0.95018,0.00000,0.00000,0.00000,0.00000
+14.00000,33.70957,0.00010,13424.76033,1.25102,0.98935,1.16827,0.99137,0.00000,0.00000,0.00000,0.00000
+15.00000,33.09508,0.00010,13400.43546,1.24817,0.97939,1.22607,0.98165,0.00000,0.00000,0.00000,0.00000
+16.00000,32.54532,0.00010,13401.86689,1.24569,0.97134,1.18299,1.00306,0.00000,0.00000,0.00000,0.00000
+17.00000,32.00117,0.00010,13416.88535,1.22134,0.96372,1.36520,0.93576,0.00000,0.00000,0.00000,0.00000
+18.00000,31.60527,0.00010,13405.81758,1.22973,0.95734,1.49343,0.97155,0.00000,0.00000,0.00000,0.00000
+19.00000,31.24081,0.00010,13381.00557,1.22577,0.95080,1.10001,0.98865,0.00000,0.00000,0.00000,0.00000
+20.00000,30.74609,0.00010,13415.91401,1.20873,0.94440,1.83207,0.92725,0.00000,0.00000,0.00000,0.00000
+21.00000,30.33131,0.00010,14342.29644,1.18136,0.93905,1.11839,0.94192,0.00000,0.00000,1.11902,0.94172
+22.00000,30.16107,0.00010,13403.45945,1.19901,0.93486,0.99529,0.88446,0.00000,0.00000,0.00000,0.00000
+23.00000,29.76894,0.00010,13384.32851,1.18600,0.92920,1.86311,0.85839,0.00000,0.00000,0.00000,0.00000
+24.00000,29.49225,0.00010,13409.76497,1.18452,0.92470,0.84768,0.89892,0.00000,0.00000,0.00000,0.00000
+25.00000,29.30378,0.00010,13426.92429,1.18931,0.92090,1.29198,0.99091,0.00000,0.00000,0.00000,0.00000
+26.00000,29.06367,0.00010,13389.06309,1.17766,0.91796,1.39764,0.89732,0.00000,0.00000,0.00000,0.00000
+27.00000,28.86095,0.00010,13437.66557,1.18131,0.91428,0.90228,0.92050,0.00000,0.00000,0.00000,0.00000
+28.00000,28.68853,0.00010,13391.12519,1.18043,0.91112,1.05935,0.92158,0.00000,0.00000,0.00000,0.00000
+29.00000,28.50810,0.00010,13418.05610,1.18622,0.90778,1.76690,0.93850,0.00000,0.00000,0.00000,0.00000
+30.00000,28.30058,0.00010,13428.80913,1.18469,0.90426,1.13567,0.91232,0.00000,0.00000,0.00000,0.00000
+31.00000,28.12497,0.00010,13418.24684,1.17267,0.90190,1.10392,0.93187,0.00000,0.00000,0.00000,0.00000
+32.00000,27.95998,0.00010,14310.87364,1.17241,0.89852,0.80569,0.88817,0.00000,0.00000,0.80583,0.88764
+33.00000,27.88917,0.00010,13402.43412,1.17662,0.89724,1.16014,0.90612,0.00000,0.00000,0.00000,0.00000
+34.00000,27.59476,0.00010,13395.82425,1.14589,0.89332,0.89120,0.88646,0.00000,0.00000,0.00000,0.00000
+35.00000,27.43520,0.00010,13397.64853,1.14857,0.89083,0.89960,0.92859,0.00000,0.00000,0.00000,0.00000
+36.00000,27.29206,0.00010,13380.16310,1.14914,0.88856,1.07694,0.85332,0.00000,0.00000,0.00000,0.00000
+37.00000,27.18069,0.00010,13384.81806,1.14949,0.88623,1.60705,0.95885,0.00000,0.00000,0.00000,0.00000
+38.00000,27.04478,0.00010,13416.52342,1.14581,0.88424,1.32012,0.96366,0.00000,0.00000,0.00000,0.00000
