+++ example/predict/training_11/log.csv
@@ -1,0 +1,29 @@
+epoch,loss(MSE),lr,time,tr_E(MAE),tr_F(MAE),val_E(MAE),val_F(MAE),irc_E(MAE),irc_F(MAE),test_E(MAE),test_F(MAE)
+1.00000,153.01514,0.00010,17663.95379,2.99919,1.89069,2.48002,1.57699,0.00000,0.00000,0.00000,0.00000
+2.00000,76.77380,0.00010,17712.46637,1.92726,1.44944,1.65086,1.39137,0.00000,0.00000,0.00000,0.00000
+3.00000,61.53846,0.00010,17762.20073,1.69372,1.31353,1.81658,1.27060,0.00000,0.00000,0.00000,0.00000
+4.00000,53.38121,0.00010,17780.38221,1.55292,1.23250,1.73074,1.18649,0.00000,0.00000,0.00000,0.00000
+5.00000,48.43426,0.00010,17762.63280,1.49047,1.17783,1.26852,1.16021,0.00000,0.00000,0.00000,0.00000
+6.00000,44.93820,0.00010,17775.84099,1.44039,1.13737,1.25100,1.11480,0.00000,0.00000,0.00000,0.00000
+7.00000,42.25734,0.00010,17779.89584,1.38173,1.10604,1.33353,1.09227,0.00000,0.00000,0.00000,0.00000
+8.00000,40.32728,0.00010,17775.94267,1.37405,1.08019,1.04579,1.06596,0.00000,0.00000,0.00000,0.00000
+9.00000,38.82867,0.00010,17766.55806,1.34644,1.06064,1.64659,0.99971,0.00000,0.00000,0.00000,0.00000
+10.00000,37.49264,0.00010,17802.44803,1.32747,1.04296,1.28855,1.05528,0.00000,0.00000,0.00000,0.00000
+11.00000,36.34106,0.00010,19054.41031,1.30239,1.02723,1.35203,0.98857,0.00000,0.00000,1.35225,0.98728
+12.00000,35.50280,0.00010,17777.49035,1.30320,1.01447,1.03688,0.99637,0.00000,0.00000,0.00000,0.00000
+13.00000,34.64530,0.00010,17805.14242,1.27835,1.00338,1.01916,1.01278,0.00000,0.00000,0.00000,0.00000
+14.00000,33.89599,0.00010,17822.79960,1.26740,0.99259,1.62258,0.93195,0.00000,0.00000,0.00000,0.00000
+15.00000,33.35402,0.00010,17804.75921,1.26460,0.98357,1.07128,0.97874,0.00000,0.00000,0.00000,0.00000
+16.00000,32.70668,0.00010,17807.70418,1.24693,0.97472,1.25269,0.97589,0.00000,0.00000,0.00000,0.00000
+17.00000,32.16591,0.00010,17785.38222,1.23198,0.96654,1.08627,0.98563,0.00000,0.00000,0.00000,0.00000
+18.00000,31.71046,0.00010,17806.16517,1.22199,0.95988,1.43194,0.91603,0.00000,0.00000,0.00000,0.00000
+19.00000,31.31898,0.00010,17803.46136,1.22178,0.95327,1.21942,0.89153,0.00000,0.00000,0.00000,0.00000
+20.00000,31.00897,0.00010,17792.42492,1.22846,0.94793,1.21975,0.95725,0.00000,0.00000,0.00000,0.00000
+21.00000,30.55717,0.00010,17809.01807,1.19894,0.94207,0.91181,0.94526,0.00000,0.00000,0.00000,0.00000
+22.00000,30.36314,0.00010,17805.84835,1.21589,0.93757,0.89989,0.94718,0.00000,0.00000,0.00000,0.00000
+23.00000,29.89978,0.00010,17798.26044,1.19204,0.93200,1.72513,0.89010,0.00000,0.00000,0.00000,0.00000
+24.00000,29.77293,0.00010,17786.56033,1.19497,0.92878,0.85290,0.92272,0.00000,0.00000,0.00000,0.00000
+25.00000,29.47523,0.00010,19114.72683,1.19673,0.92358,1.32899,0.85184,0.00000,0.00000,1.32648,0.85100
+26.00000,29.43812,0.00010,17801.80788,1.22117,0.92107,0.82695,0.87440,0.00000,0.00000,0.00000,0.00000
+27.00000,29.01932,0.00010,17792.65108,1.17861,0.91754,1.12740,0.91721,0.00000,0.00000,0.00000,0.00000
+28.00000,28.77596,0.00010,17801.67176,1.17894,0.91323,1.18897,0.86559,0.00000,0.00000,0.00000,0.00000
