--- example/predict/training_18/log.csv
@@ -1,33 +1,0 @@
-epoch,loss(MSE),lr,time,tr_E(MAE),tr_F(MAE),val_E(MAE),val_F(MAE),irc_E(MAE),irc_F(MAE),test_E(MAE),test_F(MAE)
-1.00000,153.90576,0.00010,17195.43462,2.97429,1.91097,2.34190,1.72751,0.00000,0.00000,0.00000,0.00000
-2.00000,75.98690,0.00010,17255.77645,1.85243,1.45733,3.41456,1.62018,0.00000,0.00000,0.00000,0.00000
-3.00000,60.30834,0.00010,17232.47509,1.58176,1.31525,1.86878,1.54275,0.00000,0.00000,0.00000,0.00000
-4.00000,51.93699,0.00010,17243.19797,1.43369,1.23003,1.84595,1.49335,0.00000,0.00000,0.00000,0.00000
-5.00000,46.56788,0.00010,17172.06831,1.33718,1.17053,1.76192,1.46731,0.00000,0.00000,0.00000,0.00000
-6.00000,42.69202,0.00010,17193.03334,1.26049,1.12563,2.01355,1.46175,0.00000,0.00000,0.00000,0.00000
-7.00000,39.81633,0.00010,17238.86223,1.20538,1.09043,2.13651,1.43064,0.00000,0.00000,0.00000,0.00000
-8.00000,37.52013,0.00010,17227.39473,1.16208,1.06166,1.54899,1.40896,0.00000,0.00000,0.00000,0.00000
-9.00000,35.66029,0.00010,17262.00448,1.12627,1.03731,1.48512,1.41960,0.00000,0.00000,0.00000,0.00000
-10.00000,34.14392,0.00010,17253.10463,1.10277,1.01652,1.76052,1.40882,0.00000,0.00000,0.00000,0.00000
-11.00000,32.84358,0.00010,18779.30733,1.07529,0.99875,1.65450,1.39458,0.00000,0.00000,1.80503,1.69031
-12.00000,31.68944,0.00010,17306.47072,1.05365,0.98254,1.53518,1.36648,0.00000,0.00000,0.00000,0.00000
-13.00000,30.70453,0.00010,17287.61550,1.03025,0.96798,1.56319,1.36823,0.00000,0.00000,0.00000,0.00000
-14.00000,29.86725,0.00010,17274.14400,1.02172,0.95550,1.49543,1.37376,0.00000,0.00000,0.00000,0.00000
-15.00000,29.07070,0.00010,17251.49549,0.99995,0.94388,1.50643,1.36665,0.00000,0.00000,0.00000,0.00000
-16.00000,28.36444,0.00010,17245.66855,0.98749,0.93318,1.59474,1.36314,0.00000,0.00000,0.00000,0.00000
-17.00000,27.75246,0.00010,17303.01633,0.97988,0.92341,1.49828,1.34323,0.00000,0.00000,0.00000,0.00000
-18.00000,27.18405,0.00010,17302.04491,0.97073,0.91444,1.37123,1.34109,0.00000,0.00000,0.00000,0.00000
-19.00000,26.67098,0.00010,17317.82982,0.95354,0.90623,1.40109,1.34045,0.00000,0.00000,0.00000,0.00000
-20.00000,26.17605,0.00010,17337.64526,0.94645,0.89857,1.34986,1.33060,0.00000,0.00000,0.00000,0.00000
-21.00000,25.74178,0.00010,17298.04603,0.93469,0.89129,1.44171,1.32981,0.00000,0.00000,0.00000,0.00000
-22.00000,25.32179,0.00010,17311.73173,0.93261,0.88447,1.42784,1.32902,0.00000,0.00000,0.00000,0.00000
-23.00000,24.94328,0.00010,18838.71175,0.92424,0.87843,1.34346,1.32372,0.00000,0.00000,1.54997,1.62433
-24.00000,24.60125,0.00010,17297.69654,0.91596,0.87240,2.36047,1.31177,0.00000,0.00000,0.00000,0.00000
-25.00000,24.24766,0.00010,17278.09395,0.90666,0.86666,1.39477,1.31081,0.00000,0.00000,0.00000,0.00000
-26.00000,23.93128,0.00010,17261.28924,0.90402,0.86131,1.44584,1.30841,0.00000,0.00000,0.00000,0.00000
-27.00000,23.64262,0.00010,17311.31416,0.89426,0.85614,2.24747,1.31416,0.00000,0.00000,0.00000,0.00000
-28.00000,23.37177,0.00010,17328.67562,0.89155,0.85166,1.33344,1.29799,0.00000,0.00000,0.00000,0.00000
-29.00000,23.11835,0.00010,17276.03330,0.88585,0.84711,1.63434,1.30861,0.00000,0.00000,0.00000,0.00000
-30.00000,22.89692,0.00010,17243.83098,0.88405,0.84286,1.35259,1.32408,0.00000,0.00000,0.00000,0.00000
-31.00000,22.65285,0.00010,17224.24117,0.87962,0.83875,1.64663,1.31005,0.00000,0.00000,0.00000,0.00000
-32.00000,22.42731,0.00010,17265.95838,0.87289,0.83478,1.35583,1.29513,0.00000,0.00000,0.00000,0.00000
