+++ example/predict/training_8/log.csv
@@ -1,0 +1,42 @@
+epoch,loss(MSE),lr,time,tr_E(MAE),tr_F(MAE),val_E(MAE),val_F(MAE),irc_E(MAE),irc_F(MAE),test_E(MAE),test_F(MAE)
+1.00000,152.60261,0.00010,12354.89571,2.99794,1.88964,2.04497,1.54696,0.00000,0.00000,0.00000,0.00000
+2.00000,76.74991,0.00010,12349.85252,1.92571,1.44982,2.24617,1.35185,0.00000,0.00000,0.00000,0.00000
+3.00000,61.28355,0.00010,12359.89610,1.67377,1.31220,2.50556,1.24581,0.00000,0.00000,0.00000,0.00000
+4.00000,53.17618,0.00010,12309.32777,1.54305,1.23085,1.40354,1.23405,0.00000,0.00000,0.00000,0.00000
+5.00000,48.01829,0.00010,12319.17961,1.45650,1.17463,1.19799,1.15501,0.00000,0.00000,0.00000,0.00000
+6.00000,44.38491,0.00010,12318.38717,1.39962,1.13204,1.14340,1.10756,0.00000,0.00000,0.00000,0.00000
+7.00000,41.77053,0.00010,12314.94586,1.36341,1.10029,1.23802,1.11151,0.00000,0.00000,0.00000,0.00000
+8.00000,39.71538,0.00010,12380.97632,1.32189,1.07409,1.18965,1.02331,0.00000,0.00000,0.00000,0.00000
+9.00000,38.10133,0.00010,12384.11491,1.29714,1.05304,1.26607,1.06651,0.00000,0.00000,0.00000,0.00000
+10.00000,36.76244,0.00010,12389.46686,1.27505,1.03463,1.02518,1.04583,0.00000,0.00000,0.00000,0.00000
+11.00000,35.65063,0.00010,13233.17843,1.26043,1.01925,1.01835,1.03069,0.00000,0.00000,1.01812,1.03147
+12.00000,34.67229,0.00010,12383.58693,1.24538,1.00565,1.51186,0.98168,0.00000,0.00000,0.00000,0.00000
+13.00000,33.84019,0.00010,12306.35230,1.22211,0.99406,1.19390,0.99397,0.00000,0.00000,0.00000,0.00000
+14.00000,33.24940,0.00010,12305.88672,1.22094,0.98396,1.74662,0.98290,0.00000,0.00000,0.00000,0.00000
+15.00000,32.61529,0.00010,12372.92948,1.21309,0.97453,1.33218,0.96187,0.00000,0.00000,0.00000,0.00000
+16.00000,32.07043,0.00010,12380.94106,1.20963,0.96608,1.05689,0.95507,0.00000,0.00000,0.00000,0.00000
+17.00000,31.60263,0.00010,12303.02719,1.20957,0.95832,1.18957,0.98201,0.00000,0.00000,0.00000,0.00000
+18.00000,31.08557,0.00010,12396.74435,1.18047,0.95157,1.35753,0.93117,0.00000,0.00000,0.00000,0.00000
+19.00000,30.76612,0.00010,12311.86038,1.18977,0.94527,1.12298,0.89540,0.00000,0.00000,0.00000,0.00000
+20.00000,30.32904,0.00010,12293.51132,1.17675,0.93950,0.86687,0.92166,0.00000,0.00000,0.00000,0.00000
+21.00000,30.00669,0.00010,12357.25383,1.16958,0.93426,1.17740,0.95660,0.00000,0.00000,0.00000,0.00000
+22.00000,29.69647,0.00010,13137.10946,1.17075,0.92910,0.96885,0.89188,0.00000,0.00000,0.96696,0.89284
+23.00000,29.36440,0.00010,12303.39174,1.15396,0.92463,1.23716,0.90027,0.00000,0.00000,0.00000,0.00000
+24.00000,29.11422,0.00010,12363.40968,1.15994,0.91966,0.95710,0.93537,0.00000,0.00000,0.00000,0.00000
+25.00000,28.93508,0.00010,12385.42197,1.16762,0.91591,0.89602,0.93453,0.00000,0.00000,0.00000,0.00000
+26.00000,28.57104,0.00010,12315.96098,1.13600,0.91169,1.57832,0.87479,0.00000,0.00000,0.00000,0.00000
+27.00000,28.35642,0.00010,12318.04766,1.13011,0.90802,1.13196,0.88186,0.00000,0.00000,0.00000,0.00000
+28.00000,28.06868,0.00010,12299.73372,1.12684,0.90408,1.15997,0.90809,0.00000,0.00000,0.00000,0.00000
+29.00000,27.87833,0.00010,12365.10134,1.12996,0.90057,1.31693,0.92019,0.00000,0.00000,0.00000,0.00000
+30.00000,27.67816,0.00010,12299.15339,1.12079,0.89776,1.01826,0.92260,0.00000,0.00000,0.00000,0.00000
+31.00000,27.58544,0.00010,12350.62688,1.13086,0.89492,1.21860,0.86285,0.00000,0.00000,0.00000,0.00000
+32.00000,27.33884,0.00010,12273.39308,1.11154,0.89183,0.91775,0.92191,0.00000,0.00000,0.00000,0.00000
+33.00000,27.24305,0.00010,12343.16611,1.11966,0.88939,0.90402,0.90402,0.00000,0.00000,0.00000,0.00000
+34.00000,27.03724,0.00010,13193.33408,1.10700,0.88701,0.99238,0.83562,0.00000,0.00000,0.99304,0.83640
+35.00000,26.91653,0.00010,12282.22430,1.11093,0.88453,1.37699,0.91738,0.00000,0.00000,0.00000,0.00000
+36.00000,26.75448,0.00010,12335.03880,1.10160,0.88222,0.82350,0.85813,0.00000,0.00000,0.00000,0.00000
+37.00000,26.56156,0.00010,12344.53131,1.09131,0.87984,0.88330,0.88066,0.00000,0.00000,0.00000,0.00000
+38.00000,26.54612,0.00010,12281.61458,1.10210,0.87782,1.04377,0.84378,0.00000,0.00000,0.00000,0.00000
+39.00000,26.36052,0.00010,12354.83363,1.09652,0.87593,0.81004,0.85685,0.00000,0.00000,0.00000,0.00000
+40.00000,26.24347,0.00010,12349.56587,1.08562,0.87396,1.43924,0.86029,0.00000,0.00000,0.00000,0.00000
+41.00000,26.15267,0.00010,12291.84904,1.09310,0.87207,1.42720,0.87012,0.00000,0.00000,0.00000,0.00000
