+++ example/predict/training_9/log.csv
@@ -1,0 +1,43 @@
+epoch,loss(MSE),lr,time,tr_E(MAE),tr_F(MAE),val_E(MAE),val_F(MAE),irc_E(MAE),irc_F(MAE),test_E(MAE),test_F(MAE)
+1.00000,153.29372,0.00010,12265.05374,3.00453,1.89169,2.11404,1.56764,0.00000,0.00000,0.00000,0.00000
+2.00000,76.77132,0.00010,12282.03420,1.91845,1.45021,1.90335,1.38515,0.00000,0.00000,0.00000,0.00000
+3.00000,61.31874,0.00010,12213.42327,1.67341,1.31215,1.67902,1.26566,0.00000,0.00000,0.00000,0.00000
+4.00000,53.26469,0.00010,12210.62680,1.54263,1.23148,1.24895,1.20477,0.00000,0.00000,0.00000,0.00000
+5.00000,48.03457,0.00010,12239.29625,1.45378,1.17438,2.03396,1.16922,0.00000,0.00000,0.00000,0.00000
+6.00000,44.56179,0.00010,12220.61345,1.40399,1.13351,1.39121,1.10795,0.00000,0.00000,0.00000,0.00000
+7.00000,41.92988,0.00010,12282.91374,1.36207,1.10135,1.72959,1.11269,0.00000,0.00000,0.00000,0.00000
+8.00000,39.76971,0.00010,12278.82798,1.31872,1.07485,1.22020,1.12407,0.00000,0.00000,0.00000,0.00000
+9.00000,38.22035,0.00010,12265.72917,1.29779,1.05384,1.57270,1.09365,0.00000,0.00000,0.00000,0.00000
+10.00000,36.95072,0.00010,13080.64301,1.28593,1.03628,1.52066,1.03374,0.00000,0.00000,1.52233,1.03331
+11.00000,35.83771,0.00010,12242.35800,1.26529,1.02142,1.05157,0.98642,0.00000,0.00000,0.00000,0.00000
+12.00000,34.91303,0.00010,12215.32446,1.23978,1.00828,0.91096,1.01070,0.00000,0.00000,0.00000,0.00000
+13.00000,34.09034,0.00010,12290.00130,1.22686,0.99693,1.06892,1.00358,0.00000,0.00000,0.00000,0.00000
+14.00000,33.44095,0.00010,12283.39798,1.22920,0.98619,2.43448,0.94100,0.00000,0.00000,0.00000,0.00000
+15.00000,32.85235,0.00010,12306.89717,1.21810,0.97761,0.92453,0.98173,0.00000,0.00000,0.00000,0.00000
+16.00000,32.22009,0.00010,12218.64045,1.19451,0.96920,0.96073,1.03375,0.00000,0.00000,0.00000,0.00000
+17.00000,31.70894,0.00010,12292.06873,1.18968,0.96127,1.30798,0.97186,0.00000,0.00000,0.00000,0.00000
+18.00000,31.33465,0.00010,12280.17440,1.19797,0.95432,0.86038,0.92895,0.00000,0.00000,0.00000,0.00000
+19.00000,30.84016,0.00010,12211.92902,1.18077,0.94776,0.89011,0.96175,0.00000,0.00000,0.00000,0.00000
+20.00000,30.42699,0.00010,12265.85772,1.16325,0.94188,0.90697,0.94434,0.00000,0.00000,0.00000,0.00000
+21.00000,30.10301,0.00010,12274.20102,1.16493,0.93634,1.06002,0.94840,0.00000,0.00000,0.00000,0.00000
+22.00000,29.72635,0.00010,12265.89888,1.15072,0.93087,1.08004,0.96840,0.00000,0.00000,0.00000,0.00000
+23.00000,29.46697,0.00010,13120.90529,1.15302,0.92638,0.83638,0.91460,0.00000,0.00000,0.83910,0.91438
+24.00000,29.13932,0.00010,12236.51972,1.14237,0.92151,1.08221,0.96150,0.00000,0.00000,0.00000,0.00000
+25.00000,28.85198,0.00010,12306.08986,1.12979,0.91733,1.39278,0.85849,0.00000,0.00000,0.00000,0.00000
+26.00000,28.66231,0.00010,12239.86547,1.14099,0.91348,1.62018,0.97593,0.00000,0.00000,0.00000,0.00000
+27.00000,28.41905,0.00010,12302.81759,1.13267,0.91000,1.90672,0.92174,0.00000,0.00000,0.00000,0.00000
+28.00000,28.27353,0.00010,12287.23982,1.14007,0.90603,1.00631,0.92946,0.00000,0.00000,0.00000,0.00000
+29.00000,28.05170,0.00010,12283.71217,1.12956,0.90337,0.89968,0.88104,0.00000,0.00000,0.00000,0.00000
+30.00000,27.83250,0.00010,12231.36890,1.11277,0.90007,0.83415,0.88700,0.00000,0.00000,0.00000,0.00000
+31.00000,27.71689,0.00010,12296.53999,1.12771,0.89693,1.49105,0.85942,0.00000,0.00000,0.00000,0.00000
+32.00000,27.51189,0.00010,12299.54885,1.12143,0.89394,1.82724,0.84290,0.00000,0.00000,0.00000,0.00000
+33.00000,27.29186,0.00010,12291.18661,1.10411,0.89154,1.70526,0.92367,0.00000,0.00000,0.00000,0.00000
+34.00000,27.26787,0.00010,12277.12214,1.12575,0.88911,0.87110,0.91281,0.00000,0.00000,0.00000,0.00000
+35.00000,27.05376,0.00010,13130.37045,1.10727,0.88672,0.95927,0.86479,0.00000,0.00000,0.96036,0.86471
+36.00000,26.85019,0.00010,12221.78473,1.09014,0.88439,0.87092,0.85370,0.00000,0.00000,0.00000,0.00000
+37.00000,26.79275,0.00010,12201.95804,1.10502,0.88164,0.82526,0.91409,0.00000,0.00000,0.00000,0.00000
+38.00000,26.66140,0.00010,12279.86492,1.10756,0.88016,1.22561,0.88046,0.00000,0.00000,0.00000,0.00000
+39.00000,26.52065,0.00010,12292.72134,1.09877,0.87821,0.96850,0.88971,0.00000,0.00000,0.00000,0.00000
+40.00000,26.38698,0.00010,12302.31617,1.09537,0.87563,1.44046,0.88780,0.00000,0.00000,0.00000,0.00000
+41.00000,26.32056,0.00010,12287.89405,1.08894,0.87455,0.90110,0.88477,0.00000,0.00000,0.00000,0.00000
+42.00000,26.22092,0.00010,12308.49501,1.08915,0.87220,0.94128,0.87053,0.00000,0.00000,0.00000,0.00000
